pyproject.toml
[project]
name = "wechat-adb-robot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "lxml>=5.3.0",
    "python-dotenv>=1.0.1",
    "androidviewclient>=23.0.3",
]
